Fix submitted by Martijn Pieters:
Currently, any property type int or long, set to 0, will show up in the Properties Management tab as blank. This then will result in errors when trying to change values of other properties, etc.

The problem is caused by a test for getProperty in properties.dtml. If it returns false, it won't display the value. If it would test for hasProperty instead, the property is correctly displayed.

Here is a diff:
diff -r1.29 properties.dtml
46c46
<    value="<!--#if "getProperty(id)"--><!--#var "'%d' % getProperty(id)"--><!--#/if-->">
---
>    value="<!--#if "hasProperty(id)"--><!--#var "'%d' % getProperty(id)"--><!--#/if-->">

svn path=/Zope/trunk/; revision=16405