Change the __str__() method of the HTMLParseError class to be the same as that of TALError.
Add finish_startendtag(), which works like finish_starttag() but is
called instead of that for tags of the form <foo.../>.  The default
implementation simply calls finish_starttag(), so the interface is
still compatible.

Changed comments indicating that finish_*tag() are Overridable
methods.

Got rid of the documented and implemented return value of
finish_starttag() -- it wasn't used.

svn path=/Zope/trunk/; revision=19624