Fixes so that Full now passes the basic test suite (essentially the same set of tests that Minimal also passes).  Not yet tested: versions, undo, packing, garbage collection.
InternalInconsistencyError: the base class should be AssertionError.

Fix typo in name of table that keeps the pickle reference counts (it
should be _pickleRefcounts).

next_object() => next()

Use referencef() in a completely backwards compatible way (I happen to
have an older version of ZODB installed).

load(): Fix typo in name of table that stores the pickle data (it
should be _pickles).

store(): Use a local variable called `zero' for the 8-byte zero
pointer.  When there is no current serial number for the object
(i.e. it is being stored for the first time), initialize oserial to
zero (but not orevid).  Better printing of the ConflictError message.
When there is no version string (i.e. this is being stored in a
non-version), initialize nvrevid to zero.  Better printing of the
VersionLockError.  Fix typo in return value; we should return the
cached serial number, not the serial number that was passed in as an
argument.

svn path=/Zope/trunk/; revision=19753