The doc_href method is used in two cases:
  "a link", http://an.aboslute.url

  and

  "a link":an absolute or relative URL

Fixes checked in earlier in the day to allow for only absolute URLs in the former broke regexes to do relative URLs in the latter.  This is now fixed.

FWIW, why we allow the first spelling is beyond me.  It has a tendency to be misinterpreted if it's not used with an absolute URL, which is why it's limited to matching absolute URLs currently.

svn path=/Zope/trunk/; revision=20639