Squashing unit tests failures occurring on Mac OS X.
The changes simply emulate the setup and teardown as done in the other
testing module for PageTemplates, testHTMLTests. I am not 100% sure if the fix
is the correct thing to do but it makes the tests run on Mac OS X and
does not affect them on Linux.

The main reason behind the failures is most likely some other test that runs
before which fiddles with the security policies but never cleans up
afterwards. I did some poking but could not find anything.

svn path=/Zope/trunk/; revision=21626