Merge from 2_6 branch.
Jamie Heilman discovered it was possible for anonymous users to add
ZClass permission objects.  The newly created permission objects had
no real effect on security, but anonymous users should not be able to
do this anyway, so this patch fixes the bug.

The problem was that PermissionManager never got initiatized.  While I
was here, I took the opportunity to initialize other classes as well,
update the security declaration style, and apply the 'Define
permissions' permission.

svn path=/Zope/trunk/; revision=23727