Tell subversion to expand the Id keyword:   - replaced old CVS expansion with a clean $Id$   - set the svn:keywords property to "Id"
When you check-in files in the future and want $Id$ to be expanded,
make sure it reads "$Id$" in that file; then set svn:keywords to "Id".

ZODB, ZEO, Persistence and similar packages were not modified since they
are really maintained elsewhere. Right now, we use a copy of them in our
Zope/Zope3 trees. We should go back to the module sharing (using
svn:externals).

svn path=/Zope/trunk/; revision=24763