Collector #1515: 2.8 failure: Cache statistics do not work
The DB.cacheStatistics() method was introduced in DB.py rev 1.2,
and has always returned an empty tuple.  A number of "never did
anything, probably never will" DB methods like that were removed
in ZODB 3.3, to reduce stumbling over decoys.  That was discussed
on zodb-dev at the time, with universal agreement to get rid of
them.

So the resolution in this case is changing Zope to stop trying
to access the no-longer-extant cacheStatistics() method.  If
someone doesn't like this, they can revert this patch, then
change App/CacheManager.py's cacheStatistics() to return a
hard coded empty tuple <wink>.

svn path=/Zope/trunk/; revision=27711